--- a/backend/src/events/urls.py
+++ b/backend/src/events/urls.py
@@ -1,4 +1,5 @@
 from django.conf.urls import url
+from django.views.decorators.cache import cache_page
 
 from rest_framework.urlpatterns import format_suffix_patterns
 
@@ -7,17 +8,17 @@ from .views import *
 urlpatterns = [
     url(
         regex=r"^schedule$",
-        view=ScheduleView.as_view(),
+        view=cache_page(1)(ScheduleView.as_view()),
         name="schedule_view",
     ),
     url(
         regex=r"^(?P<event_id>[0-9a-f-]+)/games$",
-        view=GamesForEventView.as_view(),
+        view=cache_page(1)(GamesForEventView.as_view()),
         name="event_games_view",
     ),
     url(
         regex=r"^locations/(?P<location_id>\d+)$",
-        view=LocationView.as_view(),
+        view=cache_page(1)(LocationView.as_view()),
         name="location_view",
     ),
 ]