--- a/backend/src/notifications/urls.py
+++ b/backend/src/notifications/urls.py
@@ -1,4 +1,5 @@
 from django.conf.urls import url
+from django.views.decorators.cache import cache_page
 
 from rest_framework.urlpatterns import format_suffix_patterns
 
@@ -7,7 +8,7 @@ from .views import *
 urlpatterns = [
     url(
         regex=r"^annoucements$",
-        view=AnnouncementView.as_view(),
+        view=cache_page(1)(AnnouncementView.as_view()),
         name="annoucement_view",
     ),
 ]