--- a/backend/src/requirements.txt
+++ b/backend/src/requirements.txt
@@ -8,3 +8,4 @@ djangorestframework==3.5.4
 whitenoise==3.3.0
 ipython==5.3.0
 drfdocs==0.0.11
+django-redis==4.7.0