--- a/backend/src/scores/urls.py
+++ b/backend/src/scores/urls.py
@@ -1,4 +1,5 @@
 from django.conf.urls import url
+from django.views.decorators.cache import cache_page
 
 from rest_framework.urlpatterns import format_suffix_patterns
 
@@ -7,12 +8,12 @@ from .views import *
 urlpatterns = [
     url(
         regex=r"^games/(?P<game_id>[0-9a-f-]+)/scores$",
-        view=ScoresForGameView.as_view(),
+        view=cache_page(1)(ScoresForGameView.as_view()),
         name="game_scores_view",
     ),
     url(
         regex=r"^games$",
-        view=GamesView.as_view(),
+        view=cache_page(1)(GamesView.as_view()),
         name="games_view",
     ),
 ]