--- a/backend/src/showdown_schedule_expanded.csv
+++ b/backend/src/showdown_schedule_expanded.csv
@@ -1,0 +1,45 @@
+﻿Welcome/Introductions,14,18,30,19,0,general,HEB University Center: Ballroom
+Modest Fashion and Henna,15,14,0,17,0,sisters,HEB University Center: Ballroom
+Halaqa,15,19,0,19,45,general,HEB University Center: Ballroom
+Performance Poetry,15,20,15,22,15,general,HEB University Center: Ballroom
+Roll Call,15,22,30,23,59,general,HEB University Center: Ballroom
+Knowledge Bowl Finals,16,8,0,9,0,general,HEB University Center: Ballroom
+Film,16,9,0,11,30,general,HEB University Center: Ballroom
+Awards,16,11,45,12,30,general,HEB University Center: Ballroom
+Registration,14,17,0,21,45,general,HEB University Center: Ballroom Galleria
+Breakfast,15,7,0,8,0,general,HEB University Center: Ballroom Galleria
+Registration,15,8,0,10,0,general,HEB University Center: Ballroom Galleria
+Dinner,15,18,45,19,0,general,HEB University Center: Ballroom Galleria
+Breakfast,16,7,30,9,0,general,HEB University Center: Ballroom Galleria
+Sisters Volleyball,15,8,0,13,30,sisters,Convocation Center
+Sr. Lunch and Dhuhr,15,13,30,14,0,sisters,Convocation Center
+Sister's Basketball,15,14,0,19,0,sisters,Convocation Center
+Sister's Dodgeball,15,14,0,19,0,sisters,Convocation Center
+Sisters Asr,15,17,5,17,15,sisters,Convocation Center
+Ice Breakers,14,19,0,19,30,general,University Center: Denman Room
+Maghrib,14,19,55,20,10,general,University Center: Denman Room
+Isha,14,22,45,22,45,general,University Center: Denman Room
+Captain Meeting,14,22,45,23,0,general,University Center: Denman Room
+Art/Scrapbook/Photography Check In,15,11,0,13,0,general,University Center: Denman Room
+Asr,15,17,5,17,15,general,University Center: Denman Room
+Maghrib,15,19,55,20,10,general,University Center: Denman Room
+Isha,15,20,15,20,30,general,University Center: Denman Room
+Dhuhr,16,13,35,13,45,general,University Center: Denman Room
+Leadership Dinner,15,19,0,19,45,general,Harris Room
+Brother's Football,15,14,0,19,0,brothers,IM Fields
+Brother's Soccer,15,14,0,19,0,brothers,IM Fields
+Brothers Asr,15,17,5,17,15,rothers,IM Fields
+Brothers Nasheed,14,20,15,22,15,brothers,Main Building: 0.106
+Sisters Nasheed,14,20,15,22,15,sisters,Main Building: 0.104
+Brothers Tajweed,14,19,0,22,15,brothers,Main Building: 1.101
+Sisters Tajweed,14,19,0,22,15,sisters,Main Building: 1.122
+Qiyam,14,22,45,23,0,genral,MCECC Masjid
+Fajr,15,6,10,6,40,general,MCECC Masjid
+Knowledge Bowl Written,15,7,0,8,0,general,MCECC Masjid
+Sack Lunches for the Homeless,15,8,0,11,30,general,MCECC Masjid
+Fajr,16,6,10,6,30,general,MCECC Masjid
+Revert Breakfast,16,7,30,8,30,general,MCECC Masjid
+FIFA Tournament,15,14,0,17,0,brothers,University Center: Mesquite Room
+Guest Speaker Discussion,15,11,45,13,30,genral,HEB University Center: Travis Room
+Brothers Basketball,15,8,0,13,30,brothers,UTSA Rec Center
+Br. Lunch and Dhuhr,15,13,30,14,0,brothers,UTSA Rec Center