… backend/src/core/migrations/0004_school_seed_data.py
… backend/src/events/migrations/0002_auto_20170407_0107.py
… backend/src/events/migrations/add_events.py