The server caches each of the client URLs just for 1 second.

Before caching: 3 requests per second (rps), 260ms mean latency
After caching: 10 requests per second (rps), 100ms mean latency

My guess is that the 5 instances of the API are causing contention with each other on the database and redis cache. Honestly was hoping for a higher RPS, but 77ms latency is awesome

WIth a longer cache timeout (120 seconds instead of 1 second), we can get 13rps and 77ms mean latency. Which means we're not sacrificing a lot of performance to cut the cache timeout down to 1 second. yay!